requirements.txt
Flask
Flask-Script
Flask-SQLAlchemy
flask-restful
flask-cors
requests
celery[redis]
psycopg2
flask-migrate
alembic
redis
validictory
